fix: fix types, sort imports, fix auth-options export
