+++ app/utils/auth.ts
@@ -2,10 +2,10 @@ import { getServerSession } from 'next-auth';
 import { StatusCodes } from 'http-status-codes';
 
 import { UserSession } from '@/common/types';
-import { authOptions } from '@/app/api/auth/[...nextauth]/route';
 import { createInnoUserIfNotExist } from '@/utils/requests/innoUsers/requests';
 
 import getLogger from './logger';
+import { authOptions } from '@/app/api/auth/[...nextauth]/options';
 
 export interface AuthResponse<TArgs> {
   status: StatusCodes;
